# backend/docs/swagger.yaml
basePath: /
definitions:
...
  http.BookRequest:
    properties:
      author:
        example: Robert C. Martin
        type: string
      isbn:
        example: "9780134494166"
        type: string
      price_cents:
        example: 3999
        type: integer
      seller_id:
        format: uuid
        type: string
      stock:
        example: 10
        type: integer
      title:
        example: Clean Architecture
        type: string
    type: object
  http.BookResponse:
    properties:
      author:
        example: Robert C. Martin
        type: string
      created_at:
        format: date-time
        type: string
      id:
        format: uuid
        type: string
      isbn:
        example: "9780134494166"
        type: string
      price_cents:
        example: 3999
        type: integer
      seller_id:
        format: uuid
        type: string
      stock:
        example: 10
        type: integer
      title:
        example: Clean Architecture
        type: string
      updated_at:
        format: date-time
        type: string
    type: object
...
  http.DeletionAcceptedResponse:
    properties:
      status:
        example: accepted
        type: string
    type: object
...
  http.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/http.ErrorDetail'
    type: object
...
  http.ModerateCommentRequest:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Public HTTP API exposed by the Book Store API Gateway.
  title: Book Store API
  version: "1.0"
paths:
  /api/v1/admin/books:
    post:
      consumes:
      - application/json
      description: Creates a book. The bearer token must contain the admin role.
      parameters:
      - description: Book payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.BookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a book
      tags:
      - Admin Books
  /api/v1/admin/books/{id}:
    delete:
      description: Deletes a book. The bearer token must contain the admin role.
      parameters:
      - description: Book ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a book
      tags:
      - Admin Books
    put:
      consumes:
      - application/json
      description: Updates a book. The bearer token must contain the admin role.
      parameters:
      - description: Book ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Book payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.BookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.BookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a book
      tags:
      - Admin Books
  /api/v1/admin/comments/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: published or hidden
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ModerateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.CommentResponse'
      security:
      - BearerAuth: []
      summary: Hide or publish a comment
      tags:
      - Admin
      - Comments
  /api/v1/admin/customers:
    get:
      description: Returns user profiles using cursor pagination. The bearer token
        must contain the admin role.
      parameters:
      - default: 20
        description: Items per request
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Opaque cursor returned by the previous request
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.CustomerListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List customers
      tags:
      - Admin Customers
  /api/v1/admin/customers/{id}:
    delete:
      description: Deletes the auth account transactionally and queues asynchronous
        profile deletion through the outbox. The bearer token must contain the admin
        role.
      parameters:
      - description: Customer ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/http.DeletionAcceptedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a customer account
      tags:
      - Admin Customers
    get:
      description: Returns one user profile by ID. The bearer token must contain the
        admin role.
      parameters:
      - description: Customer ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a customer
      tags:
      - Admin Customers
    put:
      consumes:
      - application/json
      description: Updates a customer's display name. The bearer token must contain
        the admin role.
      parameters:
      - description: Customer ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Customer profile payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a customer
      tags:
      - Admin Customers
  /api/v1/admin/wallets/{owner_id}/balance:
    put:
      consumes:
      - application/json
      description: Administrative funding/testing endpoint. It never directly overwrites
        the stored balance.
      parameters:
      - description: Wallet owner ID
        in: path
        name: owner_id
        required: true
        type: string
      - description: Unique retry key
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Signed balance delta
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateWalletBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.WalletResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Adjust a wallet balance through double-entry ledger records
      tags:
      - Admin Wallets
...
  /api/v1/auth/logout:
    post:
      description: Revokes the current refresh session and clears its HttpOnly cookie.
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Log out
      tags:
      - Auth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the token using the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"